Updated the virtualenv install instructions for windows.
With the release of virtualenv 1.10, the instructions where no longer
valid. I updated them to install the setuptools bootstrap, then
install pip anv virtualenv from pip.

I also added the installation of virtualenvwrapper-win, which is a
windows implementation of virtualenvwrapper. This makes the use of
virtualenv much easier in windows, and allow us to simplify the last
part of the guide because windows and unix instructions are now the
same.